pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pipeline"
version = "2.0.0"
description = "MUFL fantasy league data pipeline"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "supabase>=2.5",
    "psycopg2-binary>=2.9",
    "pandas>=2.2",
    "numpy>=1.26",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio", "ruff", "mypy"]
admin = ["streamlit>=1.35"]

[tool.setuptools.packages.find]
where = ["src"]
